Count the queued block updates in a long, like the other totals

+++ core/src/main/java/com/maxlananas/fawebim/core/command/RegionCommands.java
@@ -267,7 +267,7 @@ final class RegionCommands {
         entry.handler = ctx -> {
             World world = ctx.world();
             Region region = ctx.selection();
-            int updated = region.forEachPosition((x, y, z) -> {
+            long updated = region.forEachPosition((x, y, z) -> {
                 world.queueBlockUpdate(x, y, z);
                 return true;
             });
